requirements.txt
librosa==0.11.0
praat-parselmouth==0.4.6
pandas==2.3.0
numpy==1.26.4
scikit-learn==1.6.1
xgboost==3.0.2
imbalanced-learn==0.13.0
speechbrain==1.0.3
tensorflow==2.10.0
matplotlib==3.9.1
seaborn==0.13.2